efi: use size_t for UINTN array sizes instead of unsigned long
UINTN is the native register wide integer type. We don't have a specific
typedef for this in barebox, instead using size_t in places and unsigned
long in others.

The type confusion is no problem on 64-bit platforms, where size_t is
just a typedef for long. On 32-bit platforms, size_t is a typedef for
int, which while having the same size as long, is a different type,
which leads to compiler warning and errors regarding use of different
types for pointer arguments or in printf format strings.

The nice way out would have just define the size_t type as alias for
unsigned long everywhere, but this leads to false positive printf format
string warnings, so let's clean up the EFI UINTN type usage:

  - Everywhere, where the UINTN is used for an array size or buffer
    length, use size_t

  - Everywhere else keep it as unsigned long. In future, these types may
    be replaces by more specific typedefs.

Link: https://lore.barebox.org/20250527212300.575031-12-a.fatoum@barebox.org

+++ drivers/serial/serial_efi.c
@@ -75,9 +75,9 @@ struct efi_serial_io_protocol {
 	efi_status_t (EFIAPI *getcontrol) (struct efi_serial_io_protocol *This,
 			uint32_t *control);
 	efi_status_t (EFIAPI *write) (struct efi_serial_io_protocol *This,
-			unsigned long *buffersize, void *buffer);
+			size_t *buffersize, void *buffer);
 	efi_status_t (EFIAPI *read) (struct efi_serial_io_protocol *This,
-			unsigned long *buffersize, void *buffer);
+			size_t *buffersize, void *buffer);
 
 	struct efi_serial_io_mode *mode;
 };
@@ -117,7 +117,7 @@ static void efi_serial_putc(struct console_device *cdev, char c)
 	struct efi_serial_io_protocol *serial = uart->serial;
 	uint32_t control;
 	efi_status_t efiret;
-	unsigned long buffersize = sizeof(char);
+	size_t buffersize = sizeof(char);
 
 	do {
 		efiret = serial->getcontrol(serial, &control);
@@ -136,7 +136,7 @@ static int efi_serial_puts(struct console_device *cdev, const char *s,
 	struct efi_serial_io_protocol *serial = uart->serial;
 	uint32_t control;
 	efi_status_t efiret;
-	unsigned long buffersize = nbytes;
+	size_t buffersize = nbytes;
 
 	do {
 		efiret = serial->getcontrol(serial, &control);
@@ -156,7 +156,7 @@ static int efi_serial_getc(struct console_device *cdev)
 	struct efi_serial_io_protocol *serial = uart->serial;
 	uint32_t control;
 	efi_status_t efiret;
-	unsigned long buffersize = sizeof(char);
+	size_t buffersize = sizeof(char);
 	char c;
 
 	do {
